pyproject.toml
[tool.poetry]
name = "nft-image-generator"
version = "0.1.0"
description = ""
authors = ["PeteChu <phuengton.ch@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.1"
display = "^1.0.0"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
